Cleanup and remove duplicates
